### 2024-07-18 #7010 @dfm
One thing I'd like to clarify off the bat: I think that the code in the `mobse` directory is not part of this review, right? I think that this submission just covers the Python code.

Also: while this submission waits in the queue, you might want to take a look at the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) especially the [testing](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) and [installation/packaging](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions). I think that these elements look like they could use some work before the review starts!
