### 2024-07-22 b580a73
Create environment.yml
