+++ environment.yml
@@ -1,0 +1,13 @@
+name: tse
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - python=3.11.4
+  - numpy
+  - astropy
+  - scipy
+  - matplotlib
+  - pip
+  - pip:
+    - gala
