### 2024-07-22 cad748c
Update compile.sh
