+++ compile.sh
@@ -4,3 +4,5 @@ mkdir data
 cd mobse/src
 make mobse
 cd ../..
+conda env create -f environment.yml
+conda activate tse
