+++ requirements.txt
@@ -1,9 +1,0 @@
-# This file lists all the dependencies that have been tested with the current version of the code.
-Python: 3.11.4
-argparse: 1.1
-astropy: 5.1
-gala: 1.7.1
-matplotlib: 3.7.1
-numpy: 1.23.0
-scipy: 1.10.1
-sys: 3.11.4 
