### 2024-07-22 #7010 @stegmaja
@dfm - Thank you very much! Indeed, `mobse` shall not be part of the review because it is essentially copied from the MOBSE gitlab repository (https://gitlab.com/mobse/source-code). I added an instalment instruction using the new yml file and will swiftly upload a test instruction.
