### 2024-07-30 1612dbf
Update serial_batch.sh
