### 2024-08-05 #7010 @warrickball
While I continue finding reviewers, we also need to resolve the open source status of MOBSE. I note that MOBSE is MIT-licensed, which appears consistent with TSE's licence. But MOBSE appears to be derived from (and indeed has a lot of identical code to) the original BSE, which as far as I can tell was (and still is) distributed without *any* licence. In it's current state, I don't think that part of the repo can be included in our review.

There are theoretically two ways to resolve this.

1. The original author of BSE can be convinced to release it under an [OSI-approved licence](https://opensource.org/licenses) and, assuming that's the same version as used in MOBSE, MOBSE's licence can be made consistent.
2. You can remove MOBSE from the RSE repo and include its download as part of the TSE installation instructions. I think this boils down to adding `git clone https://gitlab.com/mobse/source-code mobse` before the existing instructions.

The second option is reasonably straightforward and much more feasible than the first.
