### 2024-08-08 #7010 @stegmaja
> Hi @stegmaja, I'll start looking for reviewers. If there's anyone you have in mind who isn't a [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), let me know.
>
> I'll also re-iterate @dfm's point above:
>
> > Also: while this submission waits in the queue, you might want to take a look at the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) especially the [testing](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) and [installation/packaging](https://joss.readthedocs.io/en/latest/review_criteria.html#installation-instructions). I think that these elements look like they could use some work before the review starts!
>
> I just installed `tse` and, though it worked, several points already came up:
>
> 1. The **installation** part of the `README.md` says to navigate to `mobse` to run `make mobse`. It should say "Navigate to the `mobse/src` directory." (as in `compile.sh`).
> 2. It isn't clear why it's necessary to run the `conda` steps. I didn't and the program seemed to run correctly (though without tests I can't be sure).
> 3. I couldn't find a description of the terminal output.
> 4. Nothing mentioned that a CSV file would be created in the `data` folder, nor could I find a description of its contents.

Dear @warrickball,

thank you very much for serving as an editor to my code! Regarding your first points:

1. I have changed `README.md` like you suggested.
2. The` conda` steps are to ensure that the user has all the required packages installed. While most of them are indeed standard and already installed for most users, I don't think that is the case for the `gala` package needed to (optionally) evolve the triples inside a Galactic potential.

I will add a description of 3. and 4. shortly.

Regarding your second comment about MOBSE: There are two minor changes I did to the mobse package in `mobse/src/evolve.f` and `mobse/src/mobse.f` to allow compatibility between my code and MOBSE, without actually changing the physics. Now I also highlight this in the `README.md` file. From that perspective would you think it is the most straightforward and user-friendly approach to only keep those two modified files in the repo but adding the rest of MOBSE with `git clone ...` like you suggested?

Best regards

Jakob
