### 2024-08-12 #7010 @stegmaja
Edit: 3. and 4. are also added.
