### 2024-08-14 #7010 @warrickball
Thanks for addressing these early issues. I've found two reviewers, so I'll start the proper review shortly but make a few remarks here.

> Regarding your second comment about MOBSE: There are two minor changes I did to the mobse package in mobse/src/evolve.f and mobse/src/mobse.f to allow compatibility between my code and MOBSE, without actually changing the physics. Now I also highlight this in the README.md file. From that perspective would you think it is the most straightforward and user-friendly approach to only keep those two modified files in the repo but adding the rest of MOBSE with git clone ... like you suggested?

One possibility is to include a patch file and instructions on how to apply it when installing it. You could possibly even bundle a script to download, patch and install MOBSE in the appopriate place.

To be more concrete, here's how I started a patch file that I think has the desired effect.

```
$ git clone https://github.com/stegmaja/TSE
$ git clone git@gitlab.com:mobse/source-code.git mobse
$ cd TSE
$ diff -ruN ../mobse/src mobse/src > test.patch
$ cd ../mobse
$ patch -p1 < ../TSE/test.patch
patching file src/evolve.f
patching file src/mobse.f
```
So you could include the equivalent of `test.patch` in this repo and apply it as part of the installation. It turns out you can also do something similar with `git`. In particular, replacing the last step with
```
$ git apply ../TSE/test.patch
```
worked for me.

While experimenting, I also noticed that `mobse/input/const_mobse.h` also differ. I presume that doesn't matter.

> The conda steps are to ensure that the user has all the required packages installed.

That's fair though bear in mind that not all Python users have Conda, and the implication is that TSE requires Conda, which doesn't seem to be the case.
