### 2024-08-14 #7102 @editorialbot
[REVIEW]: TSE: A triple stellar evolution code
**Submitting author:** @stegmaja (<a href="http://orcid.org/0000-0003-2340-8140">Jakob Stegmann</a>)
**Repository:** https://github.com/stegmaja/TSE
**Branch with paper.md** (empty if default branch):
**Version:** v2.0.1
