### 2024-08-14 #7102 @rieder
Hi @stegmaja, thanks for submitting!
I have a few early, small notes after reading the pre-review and doing an initial installation/run myself, more later.

- I also find Conda is not needed for the prerequisites, pip and a virtualenv would do just fine. Conda is a bit heavy as a prerequisite, so I would recommend just adding a requirements.txt for the required python packages (which can then be installed with `pip install -r requirements.txt`). In the README you could write that either a conda environment or a Python virtualenv (`python -m venv ENVIRONMENTNAME`) would work.

- Having to run `main.py` from the `./src` dir isn't great. Maybe you could have a `tse.py` script in a `./bin` directory, which calls the main function. This would help with command line history for example.

- I agree with @warrickball that it would be better to just check out the original source of MOBSE and apply a patch to it, it's best to avoid duplication.
