### 2024-08-14 783d6e8
Add a main routine to bin/ and ignore warnings in plot routine
