+++ src/main.py
@@ -10,6 +10,7 @@ import matplotlib.pyplot as plt
 from initialconditions import ic
 from spinevolution import yp_spins
 from common import dot,cross
+import warnings
 
 if ic.galactic_tides: 
     from galaxyevolution import yp_galaxy
@@ -24,6 +25,12 @@ G = G.to(u.Rsun**3/u.Msun/u.Myr**2).value
 ####################################################################################################
 
 def plot(t,y,m1,m2,m3,logr1,logr2,logr3,filename=str(ic.seed)+'_'+str(ic.Z).replace('.',''),title='Evolution of triple system'):
+
+    # Ignore all warnings within that function
+    # This is just because oftentimes we want to produce plots for certain parameters while others 
+    # (e.g., BH parameters) are not yet implemented
+    warnings.filterwarnings("ignore")
+
     print('Plotting...')
 
     
@@ -149,6 +156,9 @@ def plot(t,y,m1,m2,m3,logr1,logr2,logr3,filename=str(ic.seed)+'_'+str(ic.Z).repl
     plt.close()
 
     print('Plot saved as ./../plots/'+filename+'.png')
+
+    # No longer ignore warnings
+    warnings.filterwarnings("default")
 
 def store(t,y,star1,star2,star3,filename=str(ic.seed)+'_'+str(ic.Z).replace('.',''),status='Initial'):
 
@@ -1025,250 +1035,3 @@ def evolve(t,y,star1,star2,star3):
 
     return yp
 
-####################################################################################################
-# Main routine
-####################################################################################################
-if __name__ == '__main__':
-
-    # Initial conditions
-    m1 = ic.m1
-    m2 = ic.m2
-    m3 = ic.m3
-    a_in = ic.a1
-    a_out = ic.a2
-    e_in = ic.e1
-    e_out = ic.e2
-    cos_i_in = ic.cosi1
-    omega_in = ic.omega1
-    Omega_in = ic.Omega1
-    cos_i_out = ic.cosi2
-    omega_out = ic.omega2
-    Omega_out = ic.Omega2
-
-    lvec_in,evec_in,nvec_in = ot.orbital_elements_to_vectorial_elements(cos_i_in, Omega_in, omega_in)
-    lvec_out,evec_out,nvec_out = ot.orbital_elements_to_vectorial_elements(cos_i_out, Omega_out, omega_out)
-
-    print(ic)
-
-    # Print how the inner binary would have evolve if it were isolated
-    print('Evolve inner binary as if it was isolated',end='\n\n')
-    P_in = ot.orbital_period(ic.a1,m=ic.m1+ic.m2,units=(u.Rsun,u.day,u.Msun))
-    _ = InteractingBinaryStar(mass_1=ic.m1,mass_2=ic.m2,mass0_1=ic.m1,mass0_2=ic.m2,
-                                period=P_in,ecc=ic.e1,type1=1,type2=1,just_print=True)
-
-    # Initial state vector
-    y0 = np.zeros(26)
-
-    y0[0:3] = evec_in*e_in
-    y0[3:6] = lvec_in*np.sqrt(1-e_in**2)
-    y0[6] = a_in
-
-    y0[7:10] = evec_out*e_out
-    y0[10:13] = lvec_out*np.sqrt(1-e_out**2)
-    y0[13] = a_out
-
-    # Star objects
-    print(' ')
-    print('Evolve triple system',end='\n\n')
-    print('Evolve single stars (ignore entries for dummy secondary)',end='\n\n')
-    star1 = SingleStar(mass0_1=m1)
-    star2 = SingleStar(mass0_1=m2)
-    star3 = SingleStar(mass0_1=m3)
-
[…]
