+++ bin/tse.py
@@ -1,0 +1,261 @@
+import numpy as np
+from scipy.integrate import solve_ivp
+import sys
+import astropy.units as u
+from astropy.constants import G,c
+
+
+# Import termination events from python script located in a different directory
+sys.path.append('../src/')
+from main import plot, store, evolve_remaining_inner_binary, primary_RL, secondary_RL, tertiary_RL, unstable, primary_SN, secondary_SN, tertiary_SN, DCO_merger, Unphysical, effectively_isolated_DCO, CustomEvent, model_RLO, apply_inner_SN, apply_outer_SN, evolve
+import orbittools as ot
+from stars import SingleStar, InteractingBinaryStar
+from initialconditions import ic
+
+####################################################################################################
+# Main routine
+####################################################################################################
+if __name__ == '__main__':
+
+    # Initial conditions
+    m1 = ic.m1
+    m2 = ic.m2
+    m3 = ic.m3
+    a_in = ic.a1
+    a_out = ic.a2
+    e_in = ic.e1
+    e_out = ic.e2
+    cos_i_in = ic.cosi1
+    omega_in = ic.omega1
+    Omega_in = ic.Omega1
+    cos_i_out = ic.cosi2
+    omega_out = ic.omega2
+    Omega_out = ic.Omega2
+
+    lvec_in,evec_in,nvec_in = ot.orbital_elements_to_vectorial_elements(cos_i_in, Omega_in, omega_in)
+    lvec_out,evec_out,nvec_out = ot.orbital_elements_to_vectorial_elements(cos_i_out, Omega_out, omega_out)
+
+    print(ic)
+
+    # Print how the inner binary would have evolve if it were isolated
+    print('Evolve inner binary as if it was isolated',end='\n\n')
+    P_in = ot.orbital_period(ic.a1,m=ic.m1+ic.m2,units=(u.Rsun,u.day,u.Msun))
+    _ = InteractingBinaryStar(mass_1=ic.m1,mass_2=ic.m2,mass0_1=ic.m1,mass0_2=ic.m2,
+                                period=P_in,ecc=ic.e1,type1=1,type2=1,just_print=True)
+
+    # Initial state vector
+    y0 = np.zeros(26)
+
+    y0[0:3] = evec_in*e_in
+    y0[3:6] = lvec_in*np.sqrt(1-e_in**2)
+    y0[6] = a_in
+
+    y0[7:10] = evec_out*e_out
+    y0[10:13] = lvec_out*np.sqrt(1-e_out**2)
+    y0[13] = a_out
+
+    # Star objects
+    print(' ')
+    print('Evolve triple system',end='\n\n')
+    print('Evolve single stars (ignore entries for dummy secondary)',end='\n\n')
+    star1 = SingleStar(mass0_1=m1)
+    star2 = SingleStar(mass0_1=m2)
+    star3 = SingleStar(mass0_1=m3)
+
+    # Stellar spins
+    y0[14:17] = lvec_in*star1.interpolators['ospin'](0)*1e6
+    y0[17:20] = lvec_in*star2.interpolators['ospin'](0)*1e6
+
+    # Store initial conditions
+    store(t=0,y=y0,star1=star1,star2=star2,star3=star3,status='Initial')
+
+    # Terminating events
+    CustomEvent.terminal = True
+    primary_RL.terminal = True
+    secondary_RL.terminal = True
+    tertiary_RL.terminal = True
+    unstable.terminal = True
+    primary_SN.terminal = True
+    secondary_SN.terminal = True
+    tertiary_SN.terminal = True
+    DCO_merger.terminal = True
+    Unphysical.terminal = True
[…]
