### 2024-08-14 dc7653d
Update README.md
Update w.r.t. the new bin/tse.py file
