+++ README.md
@@ -21,17 +21,17 @@ Alternatively, simply execute `compile.sh` in the root directory. You may want t
 
 ### Running
 
-To run a system activate the environment (see above) and go to the `src` directory. Then execute
+To run a system activate the environment (see above) and go to the `bin` directory. Then execute
 
-`python main.py`.
+`python tse.py`.
 
 You may want to check out the number of optional arguments that are available by running
 
-`python main.py -h`.
+`python tse.py -h`.
 
 E.g., do 
 
-`python main.py --random TRUE --seed 1 --Z 0.02`
+`python tse.py --random TRUE --seed 1 --Z 0.02`
 
 to evolve a triple from a random distribution at solar metallicity.
 
@@ -41,7 +41,7 @@ TSE runs stably with the python version and packages listed in `requirements.txt
 
 ### Testing and output
 
-Let's test the code and inspect the output by running, e.g., `python main.py --random TRUE --seed 139 --Z 0.0001 --stellar_tides TRUE --max_time 15000.0 --method DOP853 --bhflag 1 --nsflag 3 --lamb 0.1 --alpha1 0.2`. This will lead to a terminal output similar as following.
+Let's test the code and inspect the output by running, e.g., `python tse.py --random TRUE --seed 139 --Z 0.0001 --stellar_tides TRUE --max_time 15000.0 --method DOP853 --bhflag 1 --nsflag 3 --lamb 0.1 --alpha1 0.2`. This will lead to a terminal output similar as following.
 
 First, the sampled initial conditions are printed:
 
@@ -149,19 +149,19 @@ The total evolution of the system is plotted in the directory `plots` and stored
 
 If you consider running a large population in parallel, consider using `gnu parallel`, e.g.,
 
-`parallel -j 12 "python main.py --random TRUE --seed {1} --Z {2}" ::: {1..1000} ::: 0.0002 0.002 0.02`
+`parallel -j 12 "python tse.py --random TRUE --seed {1} --Z {2}" ::: {1..1000} ::: 0.0002 0.002 0.02`
 
 which would employ `-j 12` cores to evolve 1000 systems at three different metallicities (0.0002, 0.002, 0.02) in parallel. Note that it is not supported to run other parameters than seed and metallicity in parallel. If you want to explore other parameters you need to run them one after the other, e.g.,
 
-`parallel -j 12 "python main.py --alpha1 1.0 --random TRUE --seed {1} --Z {2}" ::: {1..1000} ::: 0.0002 0.002 0.02`
+`parallel -j 12 "python tse.py --alpha1 1.0 --random TRUE --seed {1} --Z {2}" ::: {1..1000} ::: 0.0002 0.002 0.02`
 
 and then 
 
-`parallel -j 12 "python main.py --alpha1 2.0 --random TRUE --seed {1} --Z {2}" ::: {1..1000} ::: 0.0002 0.002 0.02`
+`parallel -j 12 "python tse.py --alpha1 2.0 --random TRUE --seed {1} --Z {2}" ::: {1..1000} ::: 0.0002 0.002 0.02`
 
 and **not**
 
-`parallel -j 12 "python main.py --random TRUE --seed {1} --Z {2} --alpha1 {3}" ::: {1..1000} ::: 0.0002 0.002 0.02 ::: 1.0 2.0`. 
+`parallel -j 12 "python tse.py --random TRUE --seed {1} --Z {2} --alpha1 {3}" ::: {1..1000} ::: 0.0002 0.002 0.02 ::: 1.0 2.0`. 
 
 Otherwise, different threads will use the same mobse input and output files...
 
