### 2024-08-14 988df64
Replace environment.yml by requirements.txt
