+++ requirements.txt
@@ -1,0 +1,5 @@
+numpy>=1.23.0
+astropy>=5.1
+scipy>=1.10.1
+matplotlib>=3.7.1
+gala>=1.7.1
