### 2024-09-06 #7102 @warrickball
Hi everyone! I'm just checking in on our progress here.

@stegmaja Did you make any progress with @rieder's initial comments?

@katiebreivik If you have any comments from any parts of the code and paper you have reviewed, we find it's better to post them immediately so the author can start addressing them while you have a look at other parts. You can review a few items at a time; JOSS doesn't expect monolithic reviews like mainstream journals.
