### 2024-09-09 8d690c5
Update compile.sh
