+++ compile.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 mkdir plots
 mkdir data
-cd mobse/src
+rm -rf mobse
+git clone git@gitlab.com:mobse/source-code.git mobse
+cd mobse
+patch -p1 < ../mobse.patch
+cd src
 make mobse
 cd ../..
 pip install -r requirements.txt
