### 2024-09-09 d22c5c4
Update compile.sh
