### 2024-09-09 4685020
Delete mobse directory
