+++ mobse/src/evolve.f
@@ -1,2897 +1,0 @@
-***
-      SUBROUTINE evolve(kstar,mass0,mass,rad,lumin,massc,radc,
-     &                  menv,renv,ospin,epoch,tms,
-     &                  tphys,tphysf,dtp,z,zpars,tb,ecc)
-      implicit none
-***
-*
-*           B I N A R Y
-*           ***********
-*
-*       Roche lobe overflow.
-*       --------------------
-*
-*       Developed by Jarrod Hurley, IOA, Cambridge.
-*       .........................................................
-*
-*       Advice by Christopher Tout, Onno Pols & Sverre Aarseth.
-*       ++++++++++++++++++++++++++++++++++++++++++++++++++
-*
-* Adapted from Aarseth's code 21st September 1996.
-* Fully revised on 27th November 1996 to remove vestiges of N-body code and
-* incorporate corrections.
-* Fully revised on 1st April 1998 to include new stellar evolution formulae
-* and associated binary evolution changes.
-* Fully revised on 4th July 1998 to include eccentricity, tidal 
-* circularization, wind accretion, velocity kicks for supernovae and all
-* associated orbital momentum changes.
-* Revised on 31st October 2000 to upgrade restrictions imposed on the 
-* timestep owing to magnetic braking and orbital angular momentum changes. 
-* Revised on 15th May 2016 to include a new treatment for HG stars during the CE phase and
-* velocity kick for supernova (NSs and BHs).
-*
-***
-*
-* See Tout et al., 1997, MNRAS, 291, 732 for a description of many of the
-* processes in this code as well as the relevant references mentioned
-* within the code. 
-*
-* Reference for the stellar evolution formulae is Hurley, Pols & Tout, 
-* 2000, MNRAS, 315, 543 (SSE paper).  
-* Reference for the binary evolution algorithm is Hurley, Tout & Pols, 
-* 2002, MNRAS, 329, 897 (BSE paper). 
-*
-* Reference for updates Giacobbo et al. (submitted)
-*
-***
-*
-* March 2001 *
-* Changes since version 3, i.e. since production of Paper3:  
-*
-* 1) The Eddington limit flag (on/off) has been replaced by an 
-*    Eddington limit multiplicative factor (eddfac). So if you 
-*    want to neglect the Eddington limit you would set eddfac 
-*    to a large value. 
-*
-* 2) To determine whether material transferred during RLOF forms 
-*    an accretion disk around the secondary or hits the secondary 
-*    in a direct stream we calculate a minimum radial distance, rmin, 
-*    of the mass stream from the secondary. This is taken from eq.(1) 
-*    of Ulrich & Burger (1976, ApJ, 206, 509) which they fitted to 
-*    the calculations of Lubow & Shu (1974, ApJ, 198, 383).
-*    If rmin is less than the radius of the secondary then an 
-*    accretion disk is not formed. 
-*    Note that the formula for rmin given by Ulrich & Burger is valid 
-*    for all q whereas that given by Nelemans et al. (2001, A&A, 
-*    submitted) in their eq.(6) is only valid for q < 1 where 
-*    they define q = Mdonor/Maccretor, i.e. DD systems. 
-*
-* 3) The changes to orbital and spin angular momentum owing to 
[…]
