### 2024-09-09 #7102 @stegmaja
Hi @rieder and @warrickball,

I have now addressed all your initial comments, i.e., I now also removed MOBSE from the directory but added a patch which updates the few changes files once a fresh version of MOBSE has been cloned. @warrickball Thanks for providing the steps on how to do that! All installation instructions in the Readme file have been updated accordingly.

Please let me know if there is anything else I can do at the moment.

Thank you!
