+++ paper/paper.md
@@ -10,11 +10,15 @@ tags:
 authors:
   - name: Jakob Stegmann
     orcid: 0000-0003-2340-8140
-    equal-contrib: true
     affiliation: 1 # (Multiple affiliations must be quoted)
+  - name: Fabio Antonini
+    orcid: 0000-0003-3138-6199
+    affiliation: 2 # (Multiple affiliations must be quoted)
 affiliations:
- - name: Max Planck Research Fellow, Max Planck Institute for Astrophysics, Germany
+ - name: Max Planck Research Fellow, Max Planck Institute for Astrophysics, Karl-Schwarzschild-Straße 1, 85741 Garching, Germany
    index: 1
+ - name: Gravity Exploration Institute, School of Physics and Astronomy, Cardiff University, Cardiff, CF24 3AA, UK
+   index: 2
 date: 16 July 2024
 bibliography: paper.bib
 
@@ -67,7 +71,7 @@ determine the evolution of stellar triples.
 
 # Acknowledgements
 
-I acknowledge contributions from Fabio Antonini and Jordan Barber and support from 
+I acknowledge contributions from Jordan Barber and support from 
 Michela Mapelli and Max Moe during the genesis of this project. This work was largely
 carried out at Cardiff University and was funded in parts by the Netherlands Organisation 
 for Scientific Research (NWO), as part of the Vidi research program BinWaves (project 
