### 2024-09-11 #7102 @stegmaja
Hi @warrickball,

I have reviewed contributions from different people to the code according to the JOSS's guidelines. Thus, I think Fabio Antonini should be included into the author list as he has been very active in shaping the "project direction and other forms of non-code contributions" including the physical recipe behind the code (https://joss.readthedocs.io/en/latest/submitting.html#authorship). Therefore, I have updated `paper.md` and moved him from the Acknowledgements section to the list of authors.

Best regards

Jakob
