### 2024-09-11 bdadecc
Update paper.md
Acknowledged contribution form STFC grants.
