### 2024-09-23 #7102 @katiebreivik
Hi @warrickball, @stegmaja; again, I apologize for my long delay.

I have managed to download and install TSE based on the instructions, as well as read through the paper and found no problems with either of these. I do have some suggestions to consider on the documentation and testing side:

- As far as I'm aware, all of the documentation exists in the README. I appreciate that the README contains subheadings, though I didn't immediately find them, so I'd first recommend a pointer to the subheadings.

- Ideally, it would be great to see a built out set of examples that cover the _many_ possible use cases for triple-star systems that are available from a code like TSE!  This could be completed with [readthedocs](https://docs.readthedocs.io/en/stable/tutorial/index.html) without too much fuss. At minimum though, I do think the following suggested changes to the README above will do the trick:

           1. The statement of need is currently a bit jumbled with the discussion about MOBSE and it's implementation. A non-pop-synth oriented person maybe get lost within the first few sentences so I'd suggest a clear statement of need subheading that gives a shortened version of what is already in the paper. Then a dependency subheading for the MOBSE inclusion could follow.

           2. The **Running** subheading needs some more context for what is being run with the `python tse.py` command. What should the user expect when they run the command? What does the script do without any extra command line prompts? This will then naturally lead to the second example propmt.

           3. The **Integration** variables subsection could use more context. How do I access the `y` vector? Where is it stored?

           4. Finally, I think an extra subheading that describes the output in a bit more detail (both the plotting and the stored data) would be helpful to a user. If you want to build a user base, I have found that examples for how to use the outputs is far and away the easiest way to reduce email overhead!

- For testing: when I run the tests, I do _not_ get similar expected behavior as what is shown in the README. The initial conditions start out ok, but then the output starts to differ significantly:
```Initial conditions:
 Seed=139,
 Z=0.000100,
 m1=34.868610,
 m2=31.640159,
 m3=24.505861,
 a_in=4.320278e+04,
 e_in=0.041194,
 a_out=1.726712e+06,
 e_out=0.759993,
 cos_i_in=-0.668036,
 omega_in=1.832684,
 Omega_in=3.141593,
 cos_i_out=-0.164133,
 omega_out=6.104407,
 Omega_out=0.000000

Evolve inner binary as if it was isolated

      TIME      M1       M2   K1 K2        SEP    ECC  R1/ROL1 R2/ROL2  TYPE
     0.0000   45.47890   40.52460  1  1     8671.56836  0.74   0.002   0.002  INITIAL
     4.8249   45.35591   40.44860  2  1     8691.67285  0.74   0.005   0.005  KW_CHNGE
     4.8294   45.35521   40.44833  4  1     8691.77148  0.74   0.030   0.005  KW_CHNGE
     8.1946   44.53437   40.45044  4  2     8723.55273  0.74   0.388   0.005  KW_CHNGE
     5.2000   44.51826   40.45155  4  4     8720.58008  0.74   0.389   0.024  KW_CHNGE
     5.2809   44.27627   40.45718  5  4     8673.60156  0.74   0.398   0.025  KW_CHNGE
     5.2848   44.26444   40.45819  5  4     3639.10889  0.13   1.002   0.059  BEG_RCHE
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   1.002   0.059  COMENV
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   0.000   0.223  END_RCHE
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   0.000   0.223  BEG_SYMB
     5.7805   18.57213   14.68473 14  8       14.99385  0.31   0.000   0.220  KW_CHNGE
     5.7890   18.57242   14.17974 14 14       15.14433  0.30   0.000   0.000  KW_CHNGE
   703.7903   18.57242   14.17974 14 14        0.00019  0.00   1.033   0.892  BEG_RCHE
   703.7903   19.05963   32.75216 15 14        0.00000  0.00   0.000  -1.000  COELESCE
 15000.0000    0.00000   32.75216 15 14        0.00000 -1.00  -1.000   0.000  MAX_TIME

Evolve triple system

Evolve single stars (ignore entries for dummy secondary)

      TIME      M1       M2   K1 K2        SEP    ECC  R1/ROL1 R2/ROL2  TYPE
     0.0000   45.47890   40.52460  1  1     8671.56836  0.74   0.002   0.002  INITIAL
     4.8249   45.35591   40.44860  2  1     8691.67285  0.74   0.005   0.005  KW_CHNGE
     4.8294   45.35521   40.44833  4  1     8691.77148  0.74   0.030   0.005  KW_CHNGE
     5.1946   44.53437   40.45044  4  2     8723.55273  0.74   0.388   0.005  KW_CHNGE
     5.2000   44.51826   40.45155  4  4     8720.58008  0.74   0.389   0.024  KW_CHNGE
     5.2809   44.27627   40.45718  5  4     8673.60156  0.74   0.398   0.025  KW_CHNGE
     5.2848   44.26444   40.45819  5  4     3639.10889  0.13   1.002   0.059  BEG_RCHE
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   1.002   0.059  COMENV
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   0.000   0.223  END_RCHE
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   0.000   0.223  BEG_SYMB
     5.7805   18.57213   14.68473 14  8       14.99385  0.31   0.000   0.220  KW_CHNGE
     5.7890   18.57242   14.17974 14 14       15.14433  0.30   0.000   0.000  KW_CHNGE
   703.7903   18.57242   14.17974 14 14        0.00019  0.00   1.033   0.892  BEG_RCHE
   703.7903   19.05963   32.75216 15 14        0.00000  0.00   0.000  -1.000  COELESCE
 15000.0000    0.00000   32.75216 15 14        0.00000 -1.00  -1.000   0.000  MAX_TIME

      TIME      M1       M2   K1 K2        SEP    ECC  R1/ROL1 R2/ROL2  TYPE
     0.0000   45.47890   40.52460  1  1     8671.56836  0.74   0.002   0.002  INITIAL
     4.8249   45.35591   40.44860  2  1     8691.67285  0.74   0.005   0.005  KW_CHNGE
     4.8294   45.35521   40.44833  4  1     8691.77148  0.74   0.030   0.005  KW_CHNGE
     5.1946   44.53437   40.45044  4  2     8723.55273  0.74   0.388   0.005  KW_CHNGE
     5.2000   44.51826   40.45155  4  4     8720.58008  0.74   0.389   0.024  KW_CHNGE
     5.2809   44.27627   40.45718  5  4     8673.60156  0.74   0.398   0.025  KW_CHNGE
     5.2848   44.26444   40.45819  5  4     3639.10889  0.13   1.002   0.059  BEG_RCHE
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   1.002   0.059  COMENV
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   0.000   0.223  END_RCHE
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   0.000   0.223  BEG_SYMB
     5.7805   18.57213   14.68473 14  8       14.99385  0.31   0.000   0.220  KW_CHNGE
     5.7890   18.57242   14.17974 14 14       15.14433  0.30   0.000   0.000  KW_CHNGE
   703.7903   18.57242   14.17974 14 14        0.00019  0.00   1.033   0.892  BEG_RCHE
   703.7903   19.05963   32.75216 15 14        0.00000  0.00   0.000  -1.000  COELESCE
 15000.0000    0.00000   32.75216 15 14        0.00000 -1.00  -1.000   0.000  MAX_TIME

      TIME      M1       M2   K1 K2        SEP    ECC  R1/ROL1 R2/ROL2  TYPE
     0.0000   45.47890   40.52460  1  1     8671.56836  0.74   0.002   0.002  INITIAL
     4.8249   45.35591   40.44860  2  1     8691.67285  0.74   0.005   0.005  KW_CHNGE
     4.8294   45.35521   40.44833  4  1     8691.77148  0.74   0.030   0.005  KW_CHNGE
     5.1946   44.53437   40.45044  4  2     8723.55273  0.74   0.388   0.005  KW_CHNGE
     5.2000   44.51826   40.45155  4  4     8720.58008  0.74   0.389   0.024  KW_CHNGE
     5.2809   44.27627   40.45718  5  4     8673.60156  0.74   0.398   0.025  KW_CHNGE
     5.2848   44.26444   40.45819  5  4     3639.10889  0.13   1.002   0.059  BEG_RCHE
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   1.002   0.059  COMENV
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   0.000   0.223  END_RCHE
     5.2848   18.55963   14.89280 14  7       14.92970  0.31   0.000   0.223  BEG_SYMB
     5.7805   18.57213   14.68473 14  8       14.99385  0.31   0.000   0.220  KW_CHNGE
     5.7890   18.57242   14.17974 14 14       15.14433  0.30   0.000   0.000  KW_CHNGE
   703.7903   18.57242   14.17974 14 14        0.00019  0.00   1.033   0.892  BEG_RCHE
   703.7903   19.05963   32.75216 15 14        0.00000  0.00   0.000  -1.000  COELESCE
 15000.0000    0.00000   32.75216 15 14        0.00000 -1.00  -1.000   0.000  MAX_TIME
/Users/kbreivik/Documents/Github/TSE/bin/../src/main.py:982: RuntimeWarning: invalid value encountered in sqrt
  yp[7:10] += 3./4./tLK/np.sqrt(1-e**2)*L1/L2*(dot(jv,Jv/J)*cross(Ev,jv)-5.*dot(ev,Jv/J)*cross(Ev,ev)-(1/2-3*e**2+25/2*dot(ev,Jv/J)**2-5/2*dot(jv,Jv/J)**2)*cross(Jv/J,Ev))
/Users/kbreivik/Documents/Github/TSE/bin/../src/main.py:1009: RuntimeWarning: invalid value encountered in scalar power
  yp[0:3] += -304.*ev*G**3*m1*m2*m12*(1.+121.*e**2/304.)/15./c**5/a**4/(1.-e**2)**2.5
/Users/kbreivik/Documents/Github/TSE/bin/../src/main.py:1010: RuntimeWarning: invalid value encountered in scalar power
  yp[3:6] +=  304.*e *G**3*m1*m2*m12*(1.+121.*e**2/304.)/15./c**5/a**4/(1.-e**2)**2.5*e/np.sqrt(1-e**2)*jv/j
/Users/kbreivik/Documents/Github/TSE/bin/../src/main.py:1010: RuntimeWarning: invalid value encountered in sqrt
  yp[3:6] +=  304.*e *G**3*m1*m2*m12*(1.+121.*e**2/304.)/15./c**5/a**4/(1.-e**2)**2.5*e/np.sqrt(1-e**2)*jv/j
/Users/kbreivik/Documents/Github/TSE/bin/../src/main.py:1011: RuntimeWarning: invalid value encountered in scalar power
  yp[6]   += -64.*G**3*m1*m2*m12*(1.+73.*e**2/24.+37.*e**4/96.)/5./c**5/a**3/(1.-e**2)**3.5
/Users/kbreivik/Documents/Github/TSE/bin/../src/spinevolution.py:89: RuntimeWarning: invalid value encountered in sqrt
  h1 = mu1*np.sqrt(G*m12*a)*np.sqrt(1-e**2)
/Users/kbreivik/Documents/Github/TSE/bin/../src/common.py:13: RuntimeWarning: invalid value encountered in scalar power
  return (1-e**2)**(-13/2)*(1+15/4*e**2+15/8*e**4+5/64*e**6)
/Users/kbreivik/Documents/Github/TSE/bin/../src/common.py:22: RuntimeWarning: invalid value encountered in scalar power
  return (1-e**2)**(-13/2)*(1+15/2*e**2+45/8*e**4+5/16*e**6)
A termination event occured.

Primary supernova at 8.71782017
Plotting...
Plot saved as ./../plots/139_00001.png
Inner orbit gets unbound.
```
Is it expected that I should see the same binary output repeated three times? Should a user expect to see `RuntimeWarning`'s every so often? If the answer to either of these is yes, then I think a discussion of those cases is necessary.

- Is it possible to have tests that implement random seeds such that you can show a simple test case that the user can use to verify that the code is indeed working _exactly_ as expected? This would fix the potential issues above.

- Is there functionality that is not discussed in the README? Without combing through the code, I'm finding this difficult to determine whether the functionality description is checked off; this would be solved with a readthedocs automatic API generator (though I again acknowledge that this may not be possible to implement on short timescales).

- Lastly: community guidelines are missing and should be implemented as described in the above checklist.
