+++ README.md
@@ -16,7 +16,7 @@ In order to install the code follow these steps to install
 4. Navigate to the `mobse/src` directory: `cd src`.
 5. Execute the `Makefile` by running `make mobse`. You may need to change `mobse/src/Makefile` according to the specifications of your fortran compiler.
 6. Navigate back to the root directory: `cd ./../..`.
-7. Create two directories where the outcome of the simulation is stored, `mkdir plots` and `mkdir data`.
+7. Create two directories where the outcome of the simulation is stored, `mkdir plots data`.
 8. Install all required packages by executing `pip install -r requirements.txt`.
 
 Alternatively, simply execute `install.sh` in the root directory to do all these steps at once. You may need to change the `mobse/src/Makefile` according to the specifications of your fortran compiler.
