### 2024-09-24 85e4e42
Update README.md
Describing integration variables.
