### 2024-09-25 #7102 @stegmaja
Hi @katiebreivik

thanks a lot for your thorough comments! I am about to address them shortly, but can you already tell me which python version you use? We had this issue you encountered when testing the code once when we did not use python3... so I just want to double-check that this is not the origin of the bug. I have tested on a couple of different machines now, and all yielded the output as in the Readme.md when using version 3.

Thanks!
