### 2024-09-25 #7102 @katiebreivik
Hey @stegmaja; yes sorry for not adding in my Python setup before!

I've just done a `git pull`, a fresh install, and am running on:
Python=3.10.11
Numpy=2.1.1
gcc=Apple clang version 14.0.3

After running `python tse.py`, this is my new output:

```Initial conditions:
 Seed=42,
 Z=0.020000,
 m1=1.000000,
 m2=1.000000,
 m3=1.000000,
 a_in=2.000000e+03,
 e_in=0.100000,
 a_out=2.000000e+05,
 e_out=0.500000,
 cos_i_in=1.000000,
 omega_in=0.000000,
 Omega_in=3.141593,
 cos_i_out=1.000000,
 omega_out=0.000000,
 Omega_out=0.000000

Evolve inner binary as if it was isolated

      TIME      M1       M2   K1 K2        SEP    ECC  R1/ROL1 R2/ROL2  TYPE
     0.0000    1.00000    1.00000  1  1     1999.29749  0.10   0.001   0.001  INITIAL
 11003.1299    1.00000    1.00000  2  2     1999.29749  0.10   0.002   0.002  KW_CHNGE
 11582.7891    0.99950    0.99950  3  3     2000.30688  0.10   0.003   0.003  KW_CHNGE
 12326.2852    0.76762    0.76762  4  4     2554.03711  0.10   0.011   0.011  KW_CHNGE
 12458.6348    0.72896    0.72896  5  5     2689.27075  0.10   0.023   0.023  KW_CHNGE
 12463.2559    0.65241    0.65241  6  6     2978.98755  0.10   0.202   0.202  KW_CHNGE
 12463.5068    0.52097    0.52097 11 11     3636.76392  0.10   0.000   0.000  KW_CHNGE
 15000.0000    0.52097    0.52097 11 11     3636.76392  0.10   0.000   0.000  MAX_TIME
 Fallback:   0.0000000000000000

Evolve triple system

Evolve single stars (ignore entries for dummy secondary)

      TIME      M1       M2   K1 K2        SEP    ECC  R1/ROL1 R2/ROL2  TYPE
     0.0000    1.00000    0.00100  1 14 19532744.00000  0.00   0.000   0.000  INITIAL
 11003.1299    1.00000    0.00100  2 14 19532744.00000  0.00   0.000   0.000  KW_CHNGE
 11582.7891    0.99950    0.00100  3 14 19542596.00000  0.00   0.000   0.000  KW_CHNGE
 12326.2852    0.76329    0.00100  4 14 25582410.00000  0.00   0.000   0.000  KW_CHNGE
 12458.6924    0.72467    0.00100  5 14 26943640.00000  0.00   0.000   0.000  KW_CHNGE
 12463.3184    0.64646    0.00100  6 14 30198356.00000  0.00   0.000   0.000  KW_CHNGE
 12463.5664    0.52080    0.00100 11 14 37471244.00000  0.00   0.000   0.000  KW_CHNGE
 15000.0000    0.52080    0.00100 11 14 37471244.00000  0.00   0.000   0.000  MAX_TIME
 Fallback:   0.0000000000000000

      TIME      M1       M2   K1 K2        SEP    ECC  R1/ROL1 R2/ROL2  TYPE
     0.0000    1.00000    0.00100  1 14 19532744.00000  0.00   0.000   0.000  INITIAL
 11003.1299    1.00000    0.00100  2 14 19532744.00000  0.00   0.000   0.000  KW_CHNGE
 11582.7891    0.99950    0.00100  3 14 19542596.00000  0.00   0.000   0.000  KW_CHNGE
 12326.2852    0.76329    0.00100  4 14 25582410.00000  0.00   0.000   0.000  KW_CHNGE
 12458.6924    0.72467    0.00100  5 14 26943640.00000  0.00   0.000   0.000  KW_CHNGE
 12463.3184    0.64646    0.00100  6 14 30198356.00000  0.00   0.000   0.000  KW_CHNGE
 12463.5664    0.52080    0.00100 11 14 37471244.00000  0.00   0.000   0.000  KW_CHNGE
 15000.0000    0.52080    0.00100 11 14 37471244.00000  0.00   0.000   0.000  MAX_TIME
 Fallback:   0.0000000000000000

      TIME      M1       M2   K1 K2        SEP    ECC  R1/ROL1 R2/ROL2  TYPE
     0.0000    1.00000    0.00100  1 14 19532744.00000  0.00   0.000   0.000  INITIAL
 11003.1299    1.00000    0.00100  2 14 19532744.00000  0.00   0.000   0.000  KW_CHNGE
 11582.7891    0.99950    0.00100  3 14 19542596.00000  0.00   0.000   0.000  KW_CHNGE
 12326.2852    0.76329    0.00100  4 14 25582410.00000  0.00   0.000   0.000  KW_CHNGE
 12458.6924    0.72467    0.00100  5 14 26943640.00000  0.00   0.000   0.000  KW_CHNGE
 12463.3184    0.64646    0.00100  6 14 30198356.00000  0.00   0.000   0.000  KW_CHNGE
 12463.5664    0.52080    0.00100 11 14 37471244.00000  0.00   0.000   0.000  KW_CHNGE
 15000.0000    0.52080    0.00100 11 14 37471244.00000  0.00   0.000   0.000  MAX_TIME
 Fallback:   0.0000000000000000
The solver successfully reached the end of the integration interval.
Plotting...
Plot saved as ./../plots/42_002.png
```

I think this is correct; if I run with the commands from the newly updated README, I also now get the same output. I'm not sure how exactly this works since the git pull appeared to only update the README/paper text/images.

All that's to say, I think that we are in the clear now (and I'm a bit flummoxed as to why I was getting the previous errors :-/ ).

I will take a careful look at the updated text by the end of the week to wrap up any other suggestions/checks!
