### 2024-09-26 e4d7dce
Delete docs/conf.py
