+++ docs/conf.py
@@ -1,224 +1,0 @@
-# Configuration file for the Sphinx documentation builder.
-#
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
-
-# -- Path setup --------------------------------------------------------------
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-#
-import os
-import sys
-import inspect
-import pathlib
-import datetime
-sys.path.insert(0, os.path.abspath('.'))
-
-# HACKS - credit to "https://github.com/rodluger/starry_process"
-sys.path.insert(1, os.path.dirname(os.path.abspath(__file__)))
-import hacks
-
-from configparser import ConfigParser
-conf = ConfigParser()
-
-docs_root = pathlib.Path(__file__).parent.resolve()
-conf.read([str(docs_root / '..' / 'setup.cfg')])
-setup_cfg = dict(conf.items('metadata'))
-
-# -- Project information -----------------------------------------------------
-
-project = setup_cfg['name']
-author = setup_cfg['author']
-copyright = '{0}, {1}'.format(
-    datetime.datetime.now().year, setup_cfg['author'])
-
-
-# -- General configuration ---------------------------------------------------
-
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
-extensions = [
-    "sphinx.ext.mathjax",
-    "matplotlib.sphinxext.plot_directive",
-    "nbsphinx",
-    'sphinx_design',
-    'sphinx_automodapi.automodapi',
-    'sphinx_automodapi.smart_resolver',
-    'IPython.sphinxext.ipython_console_highlighting',
-    'IPython.sphinxext.ipython_directive',
-    'numpydoc',
-    'sphinxcontrib.bibtex',
-    'sphinx.ext.intersphinx',
-    'sphinx_copybutton',
-    'sphinx.ext.linkcode',
-    'sphinx_gallery.gen_gallery',
-    'sphinx_togglebutton'
-]
-
-sphinx_gallery_conf = {
-    'examples_dirs': '../examples',   # path to your example scripts
-    'gallery_dirs': 'auto_examples',  # path to where to save gallery generated output
-    'reference_url': {
-        'sphinx_gallery': None,
-    },
-    'download_all_examples': False,
-    'show_signature': False,
-    'matplotlib_animations': True
-}
-
-
-intersphinx_mapping = {'python': ('https://docs.python.org/3', None),
-                       'matplotlib': ('https://matplotlib.org/stable', None),
-                       'numpy': ('https://numpy.org/doc/stable/', None),
-                       'scipy': ('https://docs.scipy.org/doc/scipy', None),
-                       'astropy': ('https://docs.astropy.org/en/stable', None),
-                       'pandas': ('https://pandas.pydata.org/docs/', None),
-                       'gala': ('http://gala.adrian.pw/en/latest/', None),
-                       'cosmic': ('https://cosmic-popsynth.github.io/docs/stable/', None),
-                       'pynbody': ('https://pynbody.github.io/pynbody/', None),
-                       'legwork': ('https://legwork.readthedocs.io/en/latest/', None),}
-
-bibtex_bibfiles = ['tutorials/refs.bib']
-
-# fix numpydoc autosummary
-numpydoc_show_class_members = False
-
-# use blockquotes (numpydoc>=0.8 only)
-numpydoc_use_blockquotes = True
-
-# auto-insert plot directive in examples
-numpydoc_use_plots = True
-
[…]
