+++ README.md
@@ -173,8 +173,10 @@ The total evolution of the system is plotted in the directory `plots` and stored
 In each run, TSE produces a plot stored in the directory `plots` which gives an overview of some of the most important properties of the stellar triple as a function of time. Additionally, the parameters at each noteworthy event (Roche-love overflow, supernova explosion, etc.) are stored as a csv table in the directory `data`. Both files are named after the random seed and metallicity of the simulated triple, e.g., for the [tested example system](https://github.com/stegmaja/TSE/blob/main/README.md#testing-and-terminal-output) we got two files `plots/139_00001.png` and `data/139_00001.csv` indicating our options `--seed 139` and `--Z 0.0001` (the decimal point in the file names, i.e., `0.0001` becomes `00001`). The plot is constantly updated after each noteworthy event. After the [tested example system](https://github.com/stegmaja/TSE/blob/main/README.md#testing-and-terminal-output) has finished its evolution the plot should look like:
 
 <p align="center">  
-  <img width=500 src="docs/139_00001.png">
+  <img width=1000 src="docs/139_00001.png">
 </p>
+
+The plot displays the following information. As a title we have the name of the last noteworthy event. Since the integration of that particular system was terminated after the inner orbit got unbound during the secondary supernova, the title is `Secondary supernova`. Note that this title changes while the plot is being updated during the simulation. The uppermost panels show in solid lines the semi-major axes and eccentricities of the inner (blue) and outer (orange) binary. Additionally, the dashed lines in the uppermost right panel indicate the maximum eccentricity that was ever achieved during the integration. This is sometimes helpful when the eccentricity undergoes strong oscillations and you would like to see which peak caused the eccentricity maximum (i.e., minimum of $1-e$ shown in this panel). The next row of panels shows the evolution of the stellar masses and radii. Per definition the `Primary` (blue) refers to the initially more massive star in the inner binary, `Secondary` (orange) to the initially less massive star in the inner binary, and `Tertiary` (green) to the tertiary companion in the triple. Note that during the evolution the `Primary` could generally become less massive than the `Secondary`, e.g., during a mass-transfer episode. In the right panel we also show in dashed lines the Roche-lobe radii neccessary for a star to initiate a mass-transfer episode. The next row shows the angles between the stellar spins (left) and potentially of any potentially formed black hole (right). In some more detail: ... 
 
 
 ### Parallel runs
