+++ README.md
@@ -176,7 +176,7 @@ In each run, TSE produces a plot stored in the directory `plots` which gives an 
   <img width=1000 src="docs/139_00001.png">
 </p>
 
-The plot displays the following information. As a title we have the name of the last noteworthy event. Since the integration of that particular system was terminated after the inner orbit got unbound during the secondary supernova, the title is `Secondary supernova`. Note that this title changes while the plot is being updated during the simulation. The uppermost panels show in solid lines the semi-major axes and eccentricities of the inner (blue) and outer (orange) binary. Additionally, the dashed lines in the uppermost right panel indicate the maximum eccentricity that was ever achieved during the integration. This is sometimes helpful when the eccentricity undergoes strong oscillations and you would like to see which peak caused the eccentricity maximum (i.e., minimum of $1-e$ shown in this panel). The next row of panels shows the evolution of the stellar masses and radii. Per definition the `Primary` (blue) refers to the initially more massive star in the inner binary, `Secondary` (orange) to the initially less massive star in the inner binary, and `Tertiary` (green) to the tertiary companion in the triple. Note that during the evolution the `Primary` could generally become less massive than the `Secondary`, e.g., during a mass-transfer episode. In the right panel we also show in dashed lines the Roche-lobe radii neccessary for a star to initiate a mass-transfer episode. The next row shows the angles between the stellar spins (left) and potentially of any potentially formed black hole (right). In some more detail: ... 
+The plot displays the following information. As a title we have the name of the last noteworthy event. Since the integration of that particular system was terminated after the inner orbit got unbound during the secondary supernova, the title is `Secondary supernova`. Note that this title changes while the plot is being updated during the simulation. The uppermost panels show in solid lines the semi-major axes and eccentricities of the inner (blue) and outer (orange) binary. Additionally, the dashed lines in the uppermost right panel indicate the maximum eccentricity that was ever achieved during the integration. This is sometimes helpful when the eccentricity undergoes strong oscillations and you would like to see which peak caused the eccentricity maximum (i.e., minimum of $1-e$ shown in this panel). The next row of panels shows the evolution of the stellar masses and radii. Per definition the `Primary` (blue) refers to the initially more massive star in the inner binary, `Secondary` (orange) to the initially less massive star in the inner binary, and `Tertiary` (green) to the tertiary companion in the triple. Note that during the evolution the `Primary` could generally become less massive than the `Secondary`, e.g., during a mass-transfer episode. In the right panel we also show in dashed lines the Roche-lobe radii neccessary for a star to initiate a mass-transfer episode (which does not happen in this example). The next row shows the angles between the stellar spins (left) and potentially of any potentially formed black hole (right). In some more detail: In the left panel, `Primary` (blue) refers to the cosine angle of the primary spin vector and the inner orbital angular momentum vector (i.e., $\cos i=\mathbf{\hat{S}_1}\cdot\mathbf{\hat{j}_1}$), `Secondary` (orange) to the cosine angle w.r.t. the secondary spin vector (i.e., $\cos i=\mathbf{\hat{S}_2}\cdot\mathbf{\hat{j}_1}$), and `Primary-Secondary` (green) to the cosine angle between both spins (i.e., $\cos i=\mathbf{\hat{S}_1}\cdot\mathbf{\hat{S}_2}$). The right panel shows the same properties once we have black holes. In this case, only the primary formed some after $6.4\rm Myr$. Therefore, only the blue `Primary` line is shown. Lastly, the the lowermost row shows (left) the cosine angles of the inner (blue) and outer orbital angular momentum vector (orange) w.r.t. $z$ axis of the fixed reference frame. The right panel is left blank.
 
 
 ### Parallel runs
