### 2024-09-27 #7102 @stegmaja
Hi @katiebreivik,

Although I am also slightly confused too why it works now since only README/paper and text/images seem to be updated, it is good that the code correctly recovers the example. Also thanks again for your comments. I hope I have addressed them by extending the `Readme.md` and `paper/paper.md` files. This should provide a more detailed explanation about the basic mechanism of the code, its data products, community guidelines, integration variables, and potential use cases. I do agree that having a full readthedocs documentation would be ideal and I will consider it in the future! Since I only had (very good) experiences with it as a user so far, I first have to look into its implementation though.

Let me know if you are happy with the latest changes and if there is anything else!

Thanks

Jakob
