### 2024-10-11 #7102 @rieder
One small change for the paper:
- TrES -> TRES
Apart from that (and re-generating the PDF for the paper of course), I'm happy to sign off on this, all boxes checked :).
