### 2024-10-11 7cce33d
Update paper.md
Typo
