### 2024-10-11 8d2311e
JOSS editorial tweaks to references ahead of publication
