### 2024-10-14 #7102 @warrickball
Thanks @stegmaja, we're now very close to moving to the "production" stage. We need at least two things:

* The Zenodo archive title must be the same as the JOSS paper, rather than the repo (but you can change that too). I.e., the titles should be "TSE: A triple stellar evolution code" rather than "stegmaja/TSE".
* I opened [a pull request](https://github.com/stegmaja/TSE/pull/1) to fix a few things in the paper. I'll make a second pass in the production stage but this will fix some initial rendering issues and should remove one "missing" DOI.
