### 2024-10-21 #7102 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
