### 2024-10-21 #7102 @warrickball
I've now put my AEiC hat on and am proceeding towards publication. I've ticked off a few things already but will give the paper a final closer check in the next day or two. Nearly there!
