### 2024-10-22 #7102 @warrickball
Ah, @stegmaja, the line in the BibTeX file with that title (*Multiple Stellar Evolution:* ...) has a trailing quotation that should be removed. Do you want to go ahead and make that change?

I apologise for pushing with that error. I'm not sure why I didn't catch it locally...
