### 2024-10-24 #7102 @warrickball
Thanks once again to @rieder & @katiebreivik for reviewing this submission! JOSS simply wouldn't be possible without its community of volunteers.
